refactor(locales): extract loadMessage to prevent circular dep (handles copilot's review comment)
